start port of userspace xv6-riscv to zig
I am trying to get the convert the userspace of this teaching OS
to zig, because I am incredibly bored and have a lot of stuff going
on. also a way for me to learn more about how C and Zig interops with
each other. this is the first step. I am able to get a binary compiled
and linked against the kernel and user deps of xv6 but can't yet get
it to execute. I'll dig into that later. let's go into the night!